Rename cross-unit to export-internal
