+++ ts/deck-options/index.ts
@@ -5,7 +5,8 @@
 @typescript-eslint/no-explicit-any: "off",
  */
 
-import "sveltelib/cross-unit";
+import "sveltelib/export-internal";
+
 import { getDeckOptionsInfo, DeckOptionsState } from "./lib";
 import { setupI18n, ModuleName } from "lib/i18n";
 import { checkNightMode } from "lib/nightmode";
